Use better time format
Before the time format was 12:34:567 meaning 12 minutes and 34.567
seconds. It's incorrect to separate the whole seconds and the fractional
part with a colon. Now the format 12:34.567 is used.

Also the method formatTime now only exists once in the project in
Common.java and both MainActivity.java and stats.java use it.